Refactor ExtensionFunctions.kt
